requirements.txt
a2a>=0.1.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.1.0
langchain-google-vertexai>=0.1.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.1.0
pydantic>=2.0.0
uvicorn>=0.23.0
click>=8.1.0
python-dotenv>=1.0.0
httpx>=0.25.0
fastapi>=0.104.0
requests>=2.31.0
